feature: supports to discard all local changes without untracked files (#1586)

+++ src/Commands/Discard.cs
@@ -10,38 +10,43 @@ namespace SourceGit.Commands
         /// <summary>
         ///     Discard all local changes (unstaged & staged)
         /// </summary>
-        /// <param name="repo"></param>
-        /// <param name="includeIgnored"></param>
-        /// <param name="log"></param>
-        public static async Task AllAsync(string repo, bool includeIgnored, Models.ICommandLog log)
+        public static async Task AllAsync(string repo, bool includeUntracked, bool includeIgnored, Models.ICommandLog log)
         {
-            var changes = await new QueryLocalChanges(repo).GetResultAsync().ConfigureAwait(false);
-            try
+            if (includeUntracked)
             {
-                foreach (var c in changes)
+                // Untracked paths that contains `.git` file (detached submodule) must be removed manually.
+                var changes = await new QueryLocalChanges(repo).GetResultAsync().ConfigureAwait(false);
+                try
                 {
-                    if (c.WorkTree == Models.ChangeState.Untracked ||
-                        c.WorkTree == Models.ChangeState.Added ||
-                        c.Index == Models.ChangeState.Added ||
-                        c.Index == Models.ChangeState.Renamed)
+                    foreach (var c in changes)
                     {
-                        var fullPath = Path.Combine(repo, c.Path);
-                        if (Directory.Exists(fullPath))
-                            Directory.Delete(fullPath, true);
-                        else
-                            File.Delete(fullPath);
+                        if (c.WorkTree == Models.ChangeState.Untracked ||
+                            c.WorkTree == Models.ChangeState.Added ||
+                            c.Index == Models.ChangeState.Added ||
+                            c.Index == Models.ChangeState.Renamed)
+                        {
+                            var fullPath = Path.Combine(repo, c.Path);
+                            if (Directory.Exists(fullPath))
+                                Directory.Delete(fullPath, true);
+                        }
                     }
                 }
+                catch (Exception e)
+                {
+                    App.RaiseException(repo, $"Failed to discard changes. Reason: {e.Message}");
+                }
+
+                if (includeIgnored)
+                    await new Clean(repo, Models.CleanMode.All).Use(log).ExecAsync().ConfigureAwait(false);
+                else
+                    await new Clean(repo, Models.CleanMode.OnlyUntrackedFiles).Use(log).ExecAsync().ConfigureAwait(false);
             }
-            catch (Exception e)
+            else if (includeIgnored)
             {
-                App.RaiseException(repo, $"Failed to discard changes. Reason: {e.Message}");
+                await new Clean(repo, Models.CleanMode.OnlyIgnoredFiles).Use(log).ExecAsync().ConfigureAwait(false);
             }
 
             await new Reset(repo, "HEAD", "--hard").Use(log).ExecAsync().ConfigureAwait(false);
-
-            if (includeIgnored)
-                await new Clean(repo).Use(log).ExecAsync().ConfigureAwait(false);
         }
 
         /// <summary>
